--- a/fixtures/pikachu.sh.rendered
+++ b/fixtures/pikachu.sh.rendered
@@ -1,4 +1,4 @@
-
+&nbsp;
                                                                             <span class="term-fgx52">0</span><span class="term-fgx59">101</span>
                                                                          <span class="term-fgx59">000001</span>
                                                                      <span class="term-fgx58">0</span><span class="term-fgx143">1</span><span class="term-fgx228">1</span><span class="term-fgx143">0</span><span class="term-fgx59">10001</span>
@@ -42,4 +42,5 @@
                           <span class="term-fgx58">1</span><span class="term-fgx179">0</span><span class="term-fgx221">00</span><span class="term-fgx227">100</span><span class="term-fgx221">0</span><span class="term-fgx179">0</span><span class="term-fgx136">0</span><span class="term-fgx58">0</span>                       <span class="term-fgx178">0</span><span class="term-fgx221">101</span><span class="term-fgx227">1</span><span class="term-fgx221">1</span><span class="term-fgx227">1</span><span class="term-fgx100">0</span>
                           <span class="term-fgx144">0</span><span class="term-fgx138">1</span><span class="term-fgx144">10</span><span class="term-fgx143">1</span><span class="term-fgx100">1</span><span class="term-fgx58">1</span>                            <span class="term-fgx94">0</span><span class="term-fgx222">11</span><span class="term-fgx179">1</span><span class="term-fgx228">0</span><span class="term-fgx143">0</span><span class="term-fgx180">0</span>
                                                                <span class="term-fgx59">0</span><span class="term-fgx102">1</span><span class="term-fgx144">0</span><span class="term-fgx187">0</span><span class="term-fgx101">0</span>
+&nbsp;
 &nbsp;