--- a/output.go
+++ b/output.go
@@ -66,22 +66,23 @@ func (b *outputBuffer) appendChar(char rune) {
 	}
 }
 
-func outputLineAsHTML(line screenLine) string {
+// asHTML returns the line with HTML formatting.
+func (l *screenLine) asHTML() string {
 	var spanOpen bool
 	var lineBuf outputBuffer
 
-	if data, ok := line.metadata[bkNamespace]; ok {
+	if data, ok := l.metadata[bkNamespace]; ok {
 		lineBuf.appendMeta(bkNamespace, data)
 	}
 
-	for idx, node := range line.nodes {
+	for idx, node := range l.nodes {
 		if idx == 0 {
 			if !node.style.isPlain() {
 				lineBuf.appendNodeStyle(node)
 				spanOpen = true
 			}
 		} else {
-			previous := line.nodes[idx-1]
+			previous := l.nodes[idx-1]
 			if !node.hasSameStyle(previous) {
 				if spanOpen {
 					lineBuf.closeStyle()
@@ -95,7 +96,7 @@ func outputLineAsHTML(line screenLine) string {
 		}
 
 		if node.style.element() {
-			lineBuf.buf.WriteString(line.elements[node.blob].asHTML())
+			lineBuf.buf.WriteString(l.elements[node.blob].asHTML())
 		} else {
 			lineBuf.appendChar(node.blob)
 		}
@@ -104,4 +105,17 @@ func outputLineAsHTML(line screenLine) string {
 		lineBuf.closeStyle()
 	}
 	return strings.TrimRight(lineBuf.buf.String(), " \t")
+}
+
+// asPlain returns the line contents without any added HTML.
+func (l *screenLine) asPlain() string {
+	var buf strings.Builder
+
+	for _, node := range l.nodes {
+		if !node.style.element() {
+			buf.WriteRune(node.blob)
+		}
+	}
+
+	return strings.TrimRight(buf.String(), " \t")
 }