@DrJosh9000 2024-01-18
That is the goal, to limit the memory needed to process larger and larger input. In practice it will still allocate quite a lot, but should garbage collect more of it.

Here's the same benchmark on a 17MiB file (several copies of fixtures/npm.sh.raw):

```plain
BenchmarkRendererBig-10                3         356336375 ns/op        392413698 B/op   5244420 allocs/op
BenchmarkStreamingBig-10               4         322766854 ns/op        166112616 B/op   5042078 allocs/op
```

166 MB of memory allocations per run still sounds alarming, but that number is "total" - it doesn't decrease as the garbage collector frees memory. If I stick a little diagnostic code at the end of `parseToScreen` to see what's up:

```go
	var m runtime.MemStats
	runtime.ReadMemStats(&m)
	log.Printf("TotalAlloc: %d\tHeapAlloc: %d\tHeapInuse: %d", m.TotalAlloc, m.HeapAlloc, m.HeapInuse)
```

and do a single run with the input streamed into the parser (`io.Copy`):

```plain
2024/01/18 11:26:16 TotalAlloc: 839816  HeapAlloc: 839816       HeapInuse: 1441792
2024/01/18 11:26:16 TotalAlloc: 1212656 HeapAlloc: 1212656      HeapInuse: 1761280
2024/01/18 11:26:16 TotalAlloc: 1558336 HeapAlloc: 1558336      HeapInuse: 2097152
...snip...
2024/01/18 11:26:17 TotalAlloc: 173126152       HeapAlloc: 1782400      HeapInuse: 2441216
2024/01/18 11:26:17 TotalAlloc: 173426456       HeapAlloc: 2082704      HeapInuse: 2727936
2024/01/18 11:26:17 TotalAlloc: 173618552       HeapAlloc: 2274800      HeapInuse: 2924544
```

`HeapInuse` is the most useful, and hovered ± 1.7 MB around a mean of 3.1 MB.

The same thing _without_ using `MaxLines`:

```plain
2024/01/18 11:34:43 TotalAlloc: 836776  HeapAlloc: 836776       HeapInuse: 1359872
2024/01/18 11:34:43 TotalAlloc: 1222880 HeapAlloc: 1222880      HeapInuse: 1744896
2024/01/18 11:34:43 TotalAlloc: 1527456 HeapAlloc: 1527456      HeapInuse: 2072576
...snip...
2024/01/18 11:34:43 TotalAlloc: 248025120       HeapAlloc: 154157880    HeapInuse: 166273024
2024/01/18 11:34:43 TotalAlloc: 248375520       HeapAlloc: 154508280    HeapInuse: 166453248
2024/01/18 11:34:43 TotalAlloc: 248631216       HeapAlloc: 154763976    HeapInuse: 166600704
```

i.e. heap usage grows, does not shrink, and in the end has used 166 MB without being able to free it.